requirements.txt
alembic==1.0.7
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
PyJWT==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
gunicorn==19.9.0